goods-recommendation-system-end/goods_recommendation_system/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
Django==2.2.28
django-cors-headers==3.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==2.6
jieba==0.42.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.4
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
prettytable==3.9.0
pyecharts==2.0.4
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.18.4
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.3
urllib3==1.22
wcwidth==0.2.12
